safety_report_trial/backend/requirements.txt
fastapi
uvicorn
openai
python-docx
python-multipart
langchain
faiss-cpu
sentence-transformers
sqlalchemy
passlib
bcrypt
python-jose
python-dotenv
pypdf
requests
beautifulsoup4